requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
matplotlib
wordcloud
joblib
scikit-learn
comet-ml
dvc